fix: Troop texture issues when looking at enemy troops on enemy planets

+++ Ship_Game/Data/ResourceManager.cs
@@ -637,7 +637,7 @@ namespace Ship_Game
             if (troop.StrengthMax <= 0)
                 troop.StrengthMax = troop.Strength;
 
-            troop.WhichFrame = (int)RandomMath.Float(1, troop.num_idle_frames - 1);
+            troop.WhichFrame = RandomMath.Int(1, troop.num_idle_frames - 1);
 
             if (forOwner != null)
             {
